// types/supabase.ts
export type Database = {
  public: {
    Tables: {
// ...
      conversations: {
        Row: {
          created_at: string | null
          id: string
          question: string
          response: string
          user_id: string
        }
        Insert: {
          created_at?: string | null
          id?: string
          question: string
          response: string
          user_id: string
        }
        Update: {
          created_at?: string | null
          id?: string
          question?: string
          response?: string
          user_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "public_conversations_user_id_fkey"
            columns: ["user_id"]
            isOneToOne: false
            referencedRelation: "users"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type PublicSchema = Database[Extract<keyof Database, "public">]

export type Tables<
  PublicTableNameOrOptions extends
    | keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    | { schema: keyof Database },
> = PublicTableNameOrOptions extends { schema: keyof Database }
  ? (Database[PublicTableNameOrOptions["schema"]] & {
      Tables: {
        [K in keyof (PublicSchema["Tables"] & PublicSchema["Views"])]: (PublicSchema["Tables"] & PublicSchema["Views"])[K] extends {
          Row: infer R
        }
          ? R
          : never
      }
    })[Extract<keyof (PublicSchema["Tables"] & PublicSchema["Views"]), PublicTableNameOrOptions["table"]>] extends {
      Row: infer R
    }
    ? R
    : never
  : PublicTableNameOrOptions extends keyof (PublicSchema["Tables"] & PublicSchema["Views"])
    ? (PublicSchema["Tables"] & PublicSchema["Views"])[PublicTableNameOrOptions] extends {
        Row: infer R
      }
      ? R
      : never
    : never